Tue, Jun 6, 2023, 5:30 PM +00:00

+++ ss.c
@@ -211,6 +211,10 @@ void tabulate(void)
 
 
 
+
+
+
+
 // CHECKING IF THERE IS A WINNER OR NOT
 bool print_winner(void)
 {
@@ -224,7 +228,7 @@ bool print_winner(void)
 
     for (int i = 0; i < candidate_count; i++)
     {
-        if (candidates[i].votes >= (float)total_votes / 2.0)
+        if (candidates[i].votes >= total_votes / 2.0)
         {
             printf("%s", candidates[i].name);
             winner = true;
